Split params into several *.hpp and *.cpp files
...for faster parallel compilation.
Also added apparenty forgotten files.
