+++ params/include/alps/params.hpp
@@ -19,95 +19,10 @@
 
 #include <map>
 
-#include "./params/dict_exceptions.hpp"
-#include "./params/dict_value.hpp"
+#include "./dictionary.hpp"
 
 namespace alps {
     namespace params_ns {
-
-        /// Python-like dictionary
-        class dictionary {
-          public:
-            typedef dict_value value_type;
-
-          private:
-            typedef std::map<std::string, value_type> map_type;
-            map_type map_;
-            
-          public:
-            /// Virtual destructor to make dictionary inheritable
-            virtual ~dictionary() {}
-            
-            bool empty() const { return map_.empty(); }
-            std::size_t size() const { return map_.size(); }
-
-            /// Access with intent to assign
-            value_type& operator[](const std::string& key) {
-                map_type::iterator it=map_.lower_bound(key);
-                if (it==map_.end() ||
-                    map_.key_comp()(key, it->first)) {
-                    // it's a new element, we have to construct it here
-                    // and copy it to the map, returning the ref to the inserted element
-                    it=map_.insert(it,map_type::value_type(key, value_type(key)));
-                }
-                // return reference to the existing or the newly-created element
-                return it->second;
-            }
-
-            /// Read-only access
-            const value_type& operator[](const std::string& key) const {
-                map_type::const_iterator it=map_.find(key);
-                if (it==map_.end()) throw exception::uninitialized_value(key, "Attempt to read uninitialized value");
-                return it->second;
-            }
-
-          private:
-            /// Check if the key exists and has a value; return the iterator
-            map_type::const_iterator find_nonempty_(const std::string& key) const {
-                map_type::const_iterator it=map_.find(key);
-                if (it!=map_.end() && !(it->second).empty())
-                    return it;
-                else
-                    return map_.end();
-            }
-          public:
-            
-            /// Check if a key exists and has a value (without creating the key)
-            bool exists(const std::string& key) const {
-                return find_nonempty_(key)!=map_.end();
-            }
-            
-            /// Check if a key exists and has a value of a particular type (without creating the key)
-            template <typename T>
-            bool exists(const std::string& key) const {
-                map_type::const_iterator it=find_nonempty_(key);
-                return it!=map_.end() && (it->second).isType<T>();
-            }
-
-            /// Compare two dictionaries (true if all entries are of the same type and value)
-            bool equals(const dictionary& rhs) const;
-
-            /// Save the dictionary to an archive
-            void save(alps::hdf5::archive& ar) const;
-
-            /// Load the dictionary from an archive
-            void load(alps::hdf5::archive& ar);
-
-#ifdef ALPS_HAVE_MPI
-            /// Broadcast the dictionary
-            void broadcast(const alps::mpi::communicator& comm, int root);
-#endif
-        };
-
-        inline bool operator==(const dictionary& lhs, const dictionary& rhs) {
-            return lhs.equals(rhs);
-        }
-
-        inline bool operator!=(const dictionary& lhs, const dictionary& rhs) {
-            return !(lhs==rhs);
-        }
-
-
 
         namespace detail {
             // FIXME: make it std::pair or, possibly, C++11 tuple?
@@ -258,128 +173,7 @@ namespace alps {
 
 } // alps::
 
+#include "./params/params_impl.hpp"
 
-// ** Implementation **
-#include <boost/lexical_cast.hpp>
-#include <boost/optional.hpp>
-namespace alps {
-    namespace params_ns {
-
-        namespace detail {
-            template <typename T>
-            struct parse_string {
-                static boost::optional<T> apply(const std::string& in) {
-                    T conv_result;
-                    boost::optional<T> result;
-                    if (boost::conversion::try_lexical_convert(in, conv_result)) {
-                        result=conv_result;
-                    }
-                    return result;
-                }
-            };
-
-            template <>
-            struct parse_string<std::string> {
-                static boost::optional<std::string> apply(const std::string& in) {
-                    return in;
-                }
-            };
-
-            template <>
-            struct parse_string<bool> {
-                static boost::optional<bool> apply(const std::string& in) {
-                    // FIXME: use C_locale and lowercase the string
-                    boost::optional<bool> result;
-                    if (in=="true") result=true;
-                    if (in=="false") result=false;
-                    return result;
-                }
-            };
-
-            template <typename T>
-            struct parse_string< std::vector<T> > {
-                static boost::optional< std::vector<T> > apply(const std::string& in) {
-                    typedef std::vector<T> value_type;
-                    typedef boost::optional<value_type> result_type;
-                    typedef boost::optional<T> optional_el_type;
-                    typedef std::string::const_iterator sit_type;
-                    value_type result_vec;
-                    result_type result;
-                    sit_type it1=in.begin();
-                    while (it1!=in.end()) {
-                        sit_type it2=find(it1, in.end(), ',');
-                        optional_el_type elem=parse_string<T>::apply(std::string(it1,it2));
-                        if (!elem) return result;
-                        result_vec.push_back(*elem);
-                        if (it2!=in.end()) ++it2;
-                        it1=it2;
-                    }
-                    result=result_vec;
-                    return result;
-                }
-            };
-
-        } // ::detail
-
-        template <typename T>
-        bool params::assign_to_name_(const std::string& name, const std::string& strval)
-        {
-            boost::optional<T> result=detail::parse_string<T>::apply(strval);
-            if (result) {
-                (*this)[name]=*result;
-                return true;
-            } else {
-                return false;
-            }
-        }
-        
-        template <typename T>
-        bool params::define_(const std::string& name, const std::string& descr)
-        {
-            if (this->exists(name) && !this->exists<T>(name))
-                throw exception::type_mismatch(name, "Parameter already in dictionary with a different type");
-
-            td_map_type::iterator td_it=td_map_.find(name); // FIXME: use lower-bound instead
-            if (td_it!=td_map_.end()) {
-                // FIXME: use pretty-typename!
-                if (td_it->second.typestr() != typeid(T).name()) throw exception::type_mismatch(name, "Parameter already defined with a different type");
-                td_it->second.descr()=descr;
-                return true;
-            }
-            td_map_.insert(std::make_pair(name, detail::td_pair(typeid(T).name(), descr))); // FIXME: use pretty-typename!
-
-            strmap::const_iterator it=raw_kv_content_.find(name);
-            if (it==raw_kv_content_.end()) {
-                if (this->exists(name)) return true;
-                return false; // need to decide whether the default available
-            }
-            if (!assign_to_name_<T>(name, it->second)) {
-                ++err_status_; // FIXME: record the problem: cannot parse
-                (*this)[name].clear();
-            }
-            return true;
-        }
-
-        template <typename T>
-        params& params::define(const std::string& name, const std::string& descr)
-        {
-            if (!define_<T>(name, descr)) {
-                if (!this->exists<T>(name)) ++err_status_; // FIXME: record the problem: missing required param
-            }
-            return *this;
-        }
-        
-         template <typename T>
-         params& params::define(const std::string& name, const T& defval, const std::string& descr)
-         {
-            if (!define_<T>(name, descr)) {
-                (*this)[name]=defval;
-            }
-            return *this;
-        }
-
-    } // params_ns::
-} // alps::
-        
 
 #endif /* ALPS_PARAMS_HPP_INCLUDED_00f672a032d949a7aa0e760a6b6f0602 */
